+++ crate/operator/constants.py
@@ -34,6 +34,7 @@ LABEL_USER_PASSWORD = f"operator.{API_GROUP}/user-password"
 
 SYSTEM_USERNAME = "system"
 GC_USERNAME = "gc_admin"
+GC_USER_SECRET_NAME = "user-gc-{name}"
 
 CONNECT_TIMEOUT = 10.0